requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.3
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
isort==5.13.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg2==2.9.10
python-dotenv==1.0.1
sqlparse==0.5.1
